move exported funcs from NumberInput into utils.ts; add some test cases
