+++ frontend/lib/src/components/widgets/NumberInput/NumberInput.test.tsx
@@ -27,12 +27,7 @@ import { render } from "~lib/test_util"
 import { WidgetStateManager } from "~lib/WidgetStateManager"
 import * as UseResizeObserver from "~lib/hooks/useResizeObserver"
 
-import NumberInput, {
-  canDecrement,
-  canIncrement,
-  formatValue,
-  Props,
-} from "./NumberInput"
+import NumberInput, { Props } from "./NumberInput"
 
 const getProps = (elementProps: Partial<NumberInputProto> = {}): Props => ({
   element: NumberInputProto.create({
@@ -702,84 +697,5 @@ describe("NumberInput widget", () => {
     const forId2 = numberInputLabel2.getAttribute("for")
 
     expect(forId2).toBe(forId1)
-  })
-
-  describe("utilities", () => {
-    describe("canDecrement function", () => {
-      it("returns true if decrementing stays above min", () => {
-        expect(canDecrement(5, 1, 0)).toBe(true)
-      })
-
-      it("returns false if decrementing goes below min", () => {
-        expect(canDecrement(0, 1, 0)).toBe(false)
-      })
-    })
-
-    describe("canIncrement function", () => {
-      it("returns true if incrementing stays below max", () => {
-        expect(canIncrement(5, 1, 10)).toBe(true)
-      })
-
-      it("returns false if incrementing goes above max", () => {
-        expect(canIncrement(10, 1, 10)).toBe(false)
-      })
-    })
-
-    describe("formatValue function", () => {
-      it("returns null for null value", () => {
-        expect(
-          formatValue({
-            value: null,
-            format: null,
-            step: 1,
-            dataType: NumberInputProto.DataType.INT,
-          })
-        ).toBeNull()
-      })
-
-      it("formats integer without specified format", () => {
-        expect(
-          formatValue({
-            value: 123,
-            format: null,
-            step: 1,
-            dataType: NumberInputProto.DataType.INT,
-          })
-        ).toBe("123")
-      })
-
-      it("formats float without specified format, considering step for precision", () => {
-        expect(
-          formatValue({
-            value: 123.456,
-            format: null,
-            step: 0.01,
-            dataType: NumberInputProto.DataType.FLOAT,
-          })
-        ).toBe("123.46")
-      })
-
-      it("respects format string for integers", () => {
-        expect(
-          formatValue({
-            value: 123,
-            format: "%04d",
-            step: 1,
-            dataType: NumberInputProto.DataType.INT,
-          })
-        ).toBe("0123")
-      })
-
-      it("respects format string for floats", () => {
-        expect(
-          formatValue({
-            value: 123.456,
-            format: "%.2f",
-            step: 0.01,
-            dataType: NumberInputProto.DataType.FLOAT,
-          })
-        ).toBe("123.46")
-      })
-    })
   })
 })
